+++ ChangeLog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The instrument drivers in the documentation guide are now self-contained by use of `@pyvisa-sim`` backends
 - `VISABackend._rm` is now a simple class attribute rather than a trait, so that changes to it propagate to subclasses
 - `pyvisa-sim` is now a dependency for development (but not a requirement when installing labbench as a package)
+- Cleanups in `VISABackend` and trait debug messages
 
 ### Removed
 - `SimulatedVISABackend` - instead, use `VISABackend` after calling `visa_default_resource_manager('my-file.yml@sim')`